# obviousAlexC/penPlus: optimize