@David-Orangemoon 2024-02-08
I can run pen+ on any chromium browser just fine it just seems firefox struggles.